requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
confusable-homoglyphs==3.2.0
Django==4.0.3
django-bootstrap4==22.1
django-braces==1.15.0
django-crispy-forms==1.14.0
django-model-utils==4.2.0
django-registration==3.2
django-registration-redux==2.10
django-star-ratings==0.9.2
django-tinymce==3.4.0
djangorestframework==3.13.1
Pillow==9.1.0
psycopg2==2.9.3
pytz==2022.1
soupsieve==2.3.1
sqlparse==0.4.2
swapper==1.3.0
